use HPE Greenlake header
TODO: need to add About page
