+++ src/Greeting.js
@@ -1,10 +1,7 @@
 // Greeting.js
-import React, { useContext } from 'react';
 import { PageHeader } from 'grommet';
-import { UserContext } from './global-header';
 
 const Greeting = () => {
-  const { user } = useContext(UserContext);
   return (
     <PageHeader
       title={`Welcome`}
