--- src/global-header/components/index.js
@@ -1,5 +1,0 @@
-export * from './AppIdentity';
-export * from './GlobalHeader';
-export * from './HeaderNav';
-export * from './MenuLayer';
-export * from './UserContext';
