+++ src/modal/Modal.js
@@ -1,6 +1,6 @@
-import React, {useState, useEffect} from 'react'
+import React, { useState } from 'react'
 import "./Modal.css"
-import { Card, CardHeader, Box, CardBody, CardFooter, Button, Icons, Spinner } from "grommet";
+import { Card, CardHeader, Box, CardBody, CardFooter, Button, Spinner } from "grommet";
 const Modal = ({ toggle, togglePredictive, modalToggle }) => {
   const [loader, setLoader] = useState(false);
 
@@ -10,9 +10,9 @@ const Modal = ({ toggle, togglePredictive, modalToggle }) => {
 
     // console.log("this is the loader after", loader);
     console.log("this is the loader after", loader);
-       setTimeout(() => {
-      
-       }, 2000);
+    setTimeout(() => {
+
+    }, 2000);
     togglePredictive();
     modalToggle();
   };
