+++ src/toggle/ToggleView.js
@@ -1,4 +1,4 @@
-import React, {useState, useEffect} from 'react'
+import React from 'react'
 import "./ToggleView.css"
 const ToggleView = ({ toggle,togglet}) => {
