--- src/typography/TextEmphasis.js
@@ -1,5 +1,0 @@
-import { Text } from 'grommet';
-
-export const TextEmphasis = ({ ...rest }) => {
-  return <Text weight={500} color="text-strong" {...rest} />;
-};
